feat(reporting): add multiple report formatters (console, CSV, HTML, JSON) for HTTP request reporting
