+++ reporting/formatters.go
@@ -2,11 +2,9 @@ package reporting
 
 import (
 	"bytes"
-	"encoding/csv"
 	"encoding/json"
 	"fmt"
 	"html/template"
-	"strconv"
 	"strings"
 	"time"
 
@@ -18,497 +16,9 @@ type Formatter interface {
 	Format(report *Report) (string, error)
 }
 
-// ConsoleFormatter formats reports for console output
-type ConsoleFormatter struct{}
+// HTMLFormatter.Format moved to formatter_html.go
 
-func (f *ConsoleFormatter) Format(report *Report) (string, error) {
-	var buf bytes.Buffer
-
-	buf.WriteString("HTTP Request Report\n")
-	buf.WriteString("===================\n\n")
-
-	// Summary statistics
-	buf.WriteString(fmt.Sprintf("Total Requests: %d\n", report.TotalRequests))
-	buf.WriteString(fmt.Sprintf("Successful Requests: %d\n", report.SuccessfulRequests))
-	buf.WriteString(fmt.Sprintf("Failed Requests: %d\n", report.FailedRequests))
-
-	if report.TotalRequests > 0 {
-		successRate := float64(report.SuccessfulRequests) / float64(report.TotalRequests) * 100
-		buf.WriteString(fmt.Sprintf("Success Rate: %.1f%%\n", successRate))
-	}
-
-	buf.WriteString(fmt.Sprintf("Average Response Time: %v\n", report.AverageResponseTime))
-	buf.WriteString(fmt.Sprintf("Minimum Response Time: %v\n", report.MinResponseTime))
-	buf.WriteString(fmt.Sprintf("Maximum Response Time: %v\n", report.MaxResponseTime))
-	buf.WriteString(fmt.Sprintf("Total Duration: %v\n\n", report.EndTime.Sub(report.StartTime)))
-
-	// Response time distribution
-	if len(report.ResponseTimeDistribution) > 0 {
-		buf.WriteString("Response Time Distribution:\n")
-		// Sort categories for consistent output
-		categories := []string{"<100ms", "100-500ms", "500ms-1s", ">1s"}
-		for _, category := range categories {
-			if count, exists := report.ResponseTimeDistribution[category]; exists && count > 0 {
-				buf.WriteString(fmt.Sprintf("- %s: %d\n", category, count))
-			}
-		}
-		buf.WriteString("\n")
-	}
-
-	// Error breakdown
-	if len(report.ErrorBreakdown) > 0 {
-		buf.WriteString("Error Breakdown:\n")
-		for errorType, count := range report.ErrorBreakdown {
-			buf.WriteString(fmt.Sprintf("- %s: %d\n", errorType, count))
-		}
-		buf.WriteString("\n")
-	}
-
-	// Check results summary
-	if len(report.CheckSummaries) > 0 {
-		buf.WriteString("Check Results:\n")
-		buf.WriteString(fmt.Sprintf("Total Checks: %d\n", report.TotalChecks))
-		buf.WriteString(fmt.Sprintf("Successful Checks: %d\n", report.SuccessfulChecks))
-		buf.WriteString(fmt.Sprintf("Failed Checks: %d\n", report.FailedChecks))
-
-		if report.TotalChecks > 0 {
-			checkSuccessRate := float64(report.SuccessfulChecks) / float64(report.TotalChecks) * 100
-			buf.WriteString(fmt.Sprintf("Check Success Rate: %.1f%%\n", checkSuccessRate))
-		}
-		buf.WriteString("\n")
-
-		buf.WriteString("Check Breakdown:\n")
-		for checkName, summary := range report.CheckSummaries {
-			buf.WriteString(fmt.Sprintf("- %s: %d runs (%d successful, %d failed)\n",
-				checkName, summary.TotalRuns, summary.SuccessfulRuns, summary.FailedRuns))
-			if len(summary.FailureMessages) > 0 {
-				buf.WriteString(fmt.Sprintf("  Failure reasons: %v\n", summary.FailureMessages))
-			}
-		}
-		buf.WriteString("\n")
-	}
-
-	// All collected metrics summary
-	if len(report.MetricsSummaries) > 0 {
-		buf.WriteString("All Collected Metrics:\n")
-		buf.WriteString("=====================\n\n")
-
-		// Group metrics by type for better organization
-		counters := make(map[string]metrics.MetricSummary)
-		trends := make(map[string]metrics.MetricSummary)
-		rates := make(map[string]metrics.MetricSummary)
-		gauges := make(map[string]metrics.MetricSummary)
-
-		for name, summary := range report.MetricsSummaries {
-			switch summary.Type {
-			case "counter":
-				counters[name] = summary
-			case "trend":
-				trends[name] = summary
-			case "rate":
-				rates[name] = summary
-			case "gauge":
-				gauges[name] = summary
-			}
-		}
-
-		// Display counters
-		if len(counters) > 0 {
-			buf.WriteString("Counters:\n")
-			for name, summary := range counters {
-				buf.WriteString(fmt.Sprintf("  %-25s %10.0f  (count: %d)\n", name, summary.Sum, summary.Count))
-			}
-			buf.WriteString("\n")
-		}
-
-		// Display trends (histograms)
-		if len(trends) > 0 {
-			buf.WriteString("Trends:\n")
-			for name, summary := range trends {
-				unit := "ms"
-				if name == "data_sent" || name == "data_received" {
-					unit = "bytes"
-				}
-				buf.WriteString(fmt.Sprintf("  %-25s avg=%.2f%s min=%.2f%s med=%.2f%s max=%.2f%s p(90)=%.2f%s p(95)=%.2f%s\n",
-					name, summary.Average, unit, summary.Min, unit, summary.P50, unit, summary.Max, unit, summary.P90, unit, summary.P95, unit))
-			}
-			buf.WriteString("\n")
-		}
-
-		// Display rates
-		if len(rates) > 0 {
-			buf.WriteString("Rates:\n")
-			for name, summary := range rates {
-				rate := summary.Average * 100 // Convert to percentage
-				buf.WriteString(fmt.Sprintf("  %-25s %.2f%%  (count: %d)\n", name, rate, summary.Count))
-			}
-			buf.WriteString("\n")
-		}
-
-		// Display gauges
-		if len(gauges) > 0 {
-			buf.WriteString("Gauges:\n")
-			for name, summary := range gauges {
-				buf.WriteString(fmt.Sprintf("  %-25s %.0f  (latest: %.0f)\n", name, summary.Average, summary.LatestValue))
-			}
-			buf.WriteString("\n")
-		}
-	}
-
-	// Request details (limited to first 10 for console)
-	if len(report.RequestDetails) > 0 {
-		buf.WriteString("Request Details (showing first 10):\n")
-		limit := 10
-		if len(report.RequestDetails) < limit {
-			limit = len(report.RequestDetails)
-		}
-
-		for i := 0; i < limit; i++ {
-			req := report.RequestDetails[i]
-			status := fmt.Sprintf("%d ✓", req.StatusCode)
-			if !req.Success {
-				status = fmt.Sprintf("ERROR: %s", req.Error)
-			}
-
-			// Build check failure info for this request
-			checkFailureInfo := ""
-			failedChecks := make([]string, 0)
-			for _, check := range req.Checks {
-				if !check.Success {
-					failedChecks = append(failedChecks, check.FailureMessage)
-				}
-			}
-			if len(failedChecks) > 0 {
-				checkFailureInfo = fmt.Sprintf("    %s Check Failed (Failures: %s)", req.Timestamp.Format("15:04:05.000"), strings.Join(failedChecks, ", "))
-			}
-
-			buf.WriteString(fmt.Sprintf("  %d    %s    %s    %s    %s    %.3f ms%s\n",
-				i+1, req.Name, req.Verb, req.URL, status, float64(req.ResponseTime.Nanoseconds())/1000000.0, checkFailureInfo))
-		}
-
-		if len(report.RequestDetails) > limit {
-			buf.WriteString(fmt.Sprintf("... and %d more requests\n", len(report.RequestDetails)-limit))
-		}
-	}
-
-	return buf.String(), nil
-}
-
-// HTMLFormatter formats reports as HTML
-type HTMLFormatter struct{}
-
-const htmlTemplate = `<!DOCTYPE html>
-<html>
-<head>
-    <title>HTTP Request Report</title>
-    <style>
-        body { font-family: Arial, sans-serif; margin: 20px; }
-        .summary { background: #f5f5f5; padding: 15px; border-radius: 5px; margin-bottom: 20px; }
-        .metric { display: inline-block; margin-right: 30px; }
-        table { border-collapse: collapse; width: 100%; margin-top: 20px; }
-        th, td { border: 1px solid #ddd; padding: 8px; text-align: left; }
-        th { background-color: #f2f2f2; }
-        .success { color: green; }
-        .error { color: red; }
-        .distribution, .errors { display: inline-block; vertical-align: top; margin-right: 30px; }
-        .chart { margin: 20px 0; }
-    </style>
-</head>
-<body>
-    <h1>HTTP Request Report</h1>
-    
-    <div class="summary">
-        <div class="metric"><strong>Total Requests:</strong> {{.TotalRequests}}</div>
-        <div class="metric"><strong>Successful:</strong> <span class="success">{{.SuccessfulRequests}}</span></div>
-        <div class="metric"><strong>Failed:</strong> <span class="error">{{.FailedRequests}}</span></div>
-        <div class="metric"><strong>Success Rate:</strong> {{printf "%.1f" .SuccessRate}}%</div>
-    </div>
-
-    <div class="summary">
-        <div class="metric"><strong>Avg Response Time:</strong> {{.AverageResponseTime}}</div>
-        <div class="metric"><strong>Min Response Time:</strong> {{.MinResponseTime}}</div>
-        <div class="metric"><strong>Max Response Time:</strong> {{.MaxResponseTime}}</div>
-        <div class="metric"><strong>Total Duration:</strong> {{.TotalDuration}}</div>
-    </div>
-
-    {{if .ResponseTimeDistribution}}
-    <div class="chart">
-        <div class="distribution">
-            <h3>Response Time Distribution</h3>
-            <ul>
-                {{range $category, $count := .ResponseTimeDistribution}}
-                {{if gt $count 0}}
-                <li>{{$category}}: {{$count}}</li>
-                {{end}}
-                {{end}}
-            </ul>
-        </div>
-    </div>
-    {{end}}
-
-    {{if .ErrorBreakdown}}
-    <div class="chart">
-        <div class="errors">
-            <h3>Error Breakdown</h3>
-            <ul>
-                {{range $error, $count := .ErrorBreakdown}}
-                <li>{{$error}}: {{$count}}</li>
-                {{end}}
-            </ul>
-        </div>
-    </div>
-    {{end}}
-
-    {{if .CheckSummaries}}
-    <div class="summary">
-        <div class="metric"><strong>Total Checks:</strong> {{.TotalChecks}}</div>
-        <div class="metric"><strong>Successful Checks:</strong> <span class="success">{{.SuccessfulChecks}}</span></div>
-        <div class="metric"><strong>Failed Checks:</strong> <span class="error">{{.FailedChecks}}</span></div>
-        <div class="metric"><strong>Check Success Rate:</strong> {{if gt .TotalChecks 0}}{{printf "%.1f" .CheckSuccessRate}}%{{else}}N/A{{end}}</div>
-    </div>
-
-    <div class="chart">
-        <div class="errors">
-            <h3>Check Results</h3>
-            <ul>
-                {{range $name, $summary := .CheckSummaries}}
-                <li>
-                    <strong>{{$name}}</strong>: {{$summary.TotalRuns}} runs 
-                    ({{$summary.SuccessfulRuns}} successful, {{$summary.FailedRuns}} failed)
-                    {{if gt (len $summary.FailureMessages) 0}}
-                    <br><small>Failure reasons: {{range $i, $msg := $summary.FailureMessages}}{{if $i}}, {{end}}{{$msg}}{{end}}</small>
-                    {{end}}
-                </li>
-                {{end}}
-            </ul>
-        </div>
-    </div>
-    {{end}}
-
-    {{if .MetricsSummaries}}
-    <div class="summary">
-        <h3>📊 All Collected Metrics</h3>
-        
-        <h4>Counters</h4>
-        <table style="margin-bottom: 20px;">
-            <tr><th>Metric</th><th>Value</th><th>Count</th></tr>
-            {{range $name, $summary := .MetricsSummaries}}
-                {{if eq $summary.Type "counter"}}
-                <tr><td>{{$name}}</td><td>{{printf "%.0f" $summary.Sum}}</td><td>{{$summary.Count}}</td></tr>
-                {{end}}
-            {{end}}
-        </table>
-        
-        <h4>Trends (Timing/Size Metrics)</h4>
-        <table style="margin-bottom: 20px;">
-            <tr><th>Metric</th><th>Avg</th><th>Min</th><th>Med</th><th>Max</th><th>P90</th><th>P95</th></tr>
-            {{range $name, $summary := .MetricsSummaries}}
-                {{if eq $summary.Type "trend"}}
-                <tr>
-                    <td>{{$name}}</td>
-                    <td>{{if or (eq $name "data_sent") (eq $name "data_received")}}{{printf "%.0f bytes" $summary.Average}}{{else}}{{printf "%.2f ms" $summary.Average}}{{end}}</td>
-                    <td>{{if or (eq $name "data_sent") (eq $name "data_received")}}{{printf "%.0f" $summary.Min}}{{else}}{{printf "%.2f" $summary.Min}}{{end}}</td>
-                    <td>{{if or (eq $name "data_sent") (eq $name "data_received")}}{{printf "%.0f" $summary.P50}}{{else}}{{printf "%.2f" $summary.P50}}{{end}}</td>
-                    <td>{{if or (eq $name "data_sent") (eq $name "data_received")}}{{printf "%.0f" $summary.Max}}{{else}}{{printf "%.2f" $summary.Max}}{{end}}</td>
-                    <td>{{if or (eq $name "data_sent") (eq $name "data_received")}}{{printf "%.0f" $summary.P90}}{{else}}{{printf "%.2f" $summary.P90}}{{end}}</td>
-                    <td>{{if or (eq $name "data_sent") (eq $name "data_received")}}{{printf "%.0f" $summary.P95}}{{else}}{{printf "%.2f" $summary.P95}}{{end}}</td>
-                </tr>
-                {{end}}
-            {{end}}
-        </table>
-        
-        <h4>Rates</h4>
-        <table style="margin-bottom: 20px;">
-            <tr><th>Metric</th><th>Rate</th><th>Count</th></tr>
-            {{range $name, $summary := .MetricsSummaries}}
-                {{if eq $summary.Type "rate"}}
-                <tr><td>{{$name}}</td><td>{{printf "%.2f%%" (mul $summary.Average 100.0)}}</td><td>{{$summary.Count}}</td></tr>
-                {{end}}
-            {{end}}
-        </table>
-        
-        <h4>Gauges</h4>
-        <table style="margin-bottom: 20px;">
-            <tr><th>Metric</th><th>Average</th><th>Latest</th></tr>
-            {{range $name, $summary := .MetricsSummaries}}
-                {{if eq $summary.Type "gauge"}}
-                <tr><td>{{$name}}</td><td>{{printf "%.0f" $summary.Average}}</td><td>{{printf "%.0f" $summary.LatestValue}}</td></tr>
-                {{end}}
-            {{end}}
-        </table>
-    </div>
-    {{end}}
-
-    <h3>Request Details</h3>
-    <table>
-        <tr>
-            <th>#</th>
-            <th>Name</th>
-            <th>Method</th>
-            <th>URL</th>
-            <th>Status</th>
-            <th>Response Time</th>
-            <th>Checks</th>
-            <th>Timestamp</th>
-        </tr>
-        {{range $index, $req := .RequestDetails}}
-        <tr>
-            <td>{{add $index 1}}</td>
-            <td>{{$req.Name}}</td>
-            <td>{{$req.Verb}}</td>
-            <td>{{$req.URL}}</td>
-            <td>{{if $req.Success}}<span class="success">{{$req.StatusCode}} OK</span>{{else}}<span class="error">{{$req.Error}}</span>{{end}}</td>
-            <td>{{$req.ResponseTime}}</td>
-            <td>
-                {{if gt (len $req.Checks) 0}}
-                    {{range $checkIndex, $check := $req.Checks}}
-                        {{if gt $checkIndex 0}}<br>{{end}}
-                        {{if $check.Success}}
-                            <span class="success">✓ {{$check.Name}}</span>
-                        {{else}}
-                            <span class="error">✗ {{$check.Name}}: {{$check.FailureMessage}}</span>
-                        {{end}}
-                    {{end}}
-                {{else}}
-                    -
-                {{end}}
-            </td>
-            <td>{{$req.Timestamp.Format "2006-01-02 15:04:05"}}</td>
-        </tr>
-        {{end}}
-    </table>
-</body>
-</html>`
-
-func (f *HTMLFormatter) Format(report *Report) (string, error) {
-	tmpl, err := template.New("report").Funcs(template.FuncMap{
-		"add": func(a, b int) int { return a + b },
-		"mul": func(a, b float64) float64 { return a * b },
-	}).Parse(htmlTemplate)
-	if err != nil {
-		return "", err
-	}
-
-	// Prepare template data
-	data := struct {
-		*Report
-		SuccessRate      float64
-		CheckSuccessRate float64
-		TotalDuration    time.Duration
-	}{
-		Report:        report,
-		TotalDuration: report.EndTime.Sub(report.StartTime),
-	}
-
-	if report.TotalRequests > 0 {
-		data.SuccessRate = float64(report.SuccessfulRequests) / float64(report.TotalRequests) * 100
-	}
-
-	if report.TotalChecks > 0 {
-		data.CheckSuccessRate = float64(report.SuccessfulChecks) / float64(report.TotalChecks) * 100
-	}
-
-	var buf bytes.Buffer
-	if err := tmpl.Execute(&buf, data); err != nil {
-		return "", err
-	}
-
-	return buf.String(), nil
-}
-
-// CSVFormatter formats reports as CSV
-type CSVFormatter struct{}
-
-func (f *CSVFormatter) Format(report *Report) (string, error) {
-	var buf bytes.Buffer
-	writer := csv.NewWriter(&buf)
-
-	// Write header
-	header := []string{"Index", "Name", "Method", "URL", "Success", "StatusCode", "ResponseTime", "Error", "CheckFailures", "Timestamp"}
-	if err := writer.Write(header); err != nil {
-		return "", err
-	}
-
-	// Write data
-	for i, req := range report.RequestDetails {
-		// Collect failed check messages
-		var failedChecks []string
-		for _, check := range req.Checks {
-			if !check.Success && check.FailureMessage != "" {
-				failedChecks = append(failedChecks, check.FailureMessage)
-			}
-		}
-		checkFailuresStr := strings.Join(failedChecks, "; ")
-
-		record := []string{
-			strconv.Itoa(i + 1),
-			req.Name,
-			req.Verb,
-			req.URL,
-			strconv.FormatBool(req.Success),
-			strconv.Itoa(req.StatusCode),
-			req.ResponseTime.String(),
-			req.Error,
-			checkFailuresStr,
-			req.Timestamp.Format(time.RFC3339),
-		}
-		if err := writer.Write(record); err != nil {
-			return "", err
-		}
-	}
-
-	writer.Flush()
-	if err := writer.Error(); err != nil {
-		return "", err
-	}
-
-	return buf.String(), nil
-}
-
-// JSONFormatter formats reports as JSON
-type JSONFormatter struct{}
-
-func (f *JSONFormatter) Format(report *Report) (string, error) {
-	// Create a more structured JSON output
-	output := map[string]interface{}{
-		"summary": map[string]interface{}{
-			"totalRequests":       report.TotalRequests,
-			"successfulRequests":  report.SuccessfulRequests,
-			"failedRequests":      report.FailedRequests,
-			"successRate":         0.0,
-			"totalChecks":         report.TotalChecks,
-			"successfulChecks":    report.SuccessfulChecks,
-			"failedChecks":        report.FailedChecks,
-			"checkSuccessRate":    0.0,
-			"averageResponseTime": report.AverageResponseTime.String(),
-			"minResponseTime":     report.MinResponseTime.String(),
-			"maxResponseTime":     report.MaxResponseTime.String(),
-			"totalDuration":       report.EndTime.Sub(report.StartTime).String(),
-			"startTime":           report.StartTime.Format(time.RFC3339),
-			"endTime":             report.EndTime.Format(time.RFC3339),
-		},
-		"responseTimeDistribution": report.ResponseTimeDistribution,
-		"errorBreakdown":           report.ErrorBreakdown,
-		"checkSummaries":           report.CheckSummaries,
-		"metricsSummaries":         report.MetricsSummaries,
-		"requestDetails":           report.RequestDetails,
-	}
-
-	if report.TotalRequests > 0 {
-		output["summary"].(map[string]interface{})["successRate"] = float64(report.SuccessfulRequests) / float64(report.TotalRequests) * 100
-	}
-
-	if report.TotalChecks > 0 {
-		output["summary"].(map[string]interface{})["checkSuccessRate"] = float64(report.SuccessfulChecks) / float64(report.TotalChecks) * 100
-	}
-
-	jsonBytes, err := json.MarshalIndent(output, "", "  ")
-	if err != nil {
-		return "", err
-	}
-
-	return string(jsonBytes), nil
-}
+// JSON formatter moved
 
 // GetFormatter returns the appropriate formatter for the given format
 func GetFormatter(format ReportFormat) Formatter {
